requirements.txt
fastapi
uvicorn
numpy
pillow
python-multipart
xhtml2pdf
qrcode
httpx
